代码优化；增加对已删除的函数、询问并选择的支持

+++ Main.py
@@ -1,21 +1,24 @@
 import json
-import tkinter
-from tkinter import filedialog
 import traceback
-from CommunityAPI import getCompiledWork, getWorkInfo
+
+from CommunityAPI import *
 from Decompiler import WorkDecompiler
 from MetaData import *
 from Tool import showError
 from UI import *
+from HTTP import HTTP
 
 def main():
     printMetaData()
 
     workID = UI.askInteger("请输入作品 ID")
-    workInfo = getWorkInfo(workID)
+
+    workInfo = HTTP.getJSON(API_GET_RELEASED_WORK_INFO + str(workID))
     log(INFO, f"成功获取作品 \033[4;32m{workInfo['name']}\033[0m 的信息。")
-    compiledWork = getCompiledWork(workInfo["source_urls"][0])
+
+    compiledWork = HTTP.getJSON(workInfo["source_urls"][0])
     log(INFO, f"成功获取作品 \033[4;32m{workInfo['name']}\033[0m 的编译文件。")
+
     try:
         decompiler = WorkDecompiler(workInfo, compiledWork)
         decompiler.onStart = lambda: log(INFO, f"开始反编译，作品名称：\033[4;32m{workInfo['name']}\033[0m。")
@@ -37,19 +40,24 @@ def main():
 
 def saveSourceCode(sourceCode):
     input("按下回车键保存源码")
-    root = tkinter.Tk()
-    root.title("保存源码文件")
     while True:
-        path = filedialog.asksaveasfilename(defaultextension = ".bcm", filetypes = [("源码文件", ".bcm"), ("源码4文件", ".bcm4"), ("所有文件", ".*")])
-        print(f"正在将源码保存到文件 \033[4m{path}\033[0m ……")
-        try:
-            with open(path, "w") as file:
-                json.dump(sourceCode, file)
-            break
-        except FileNotFoundError:
-            print("\033[1;7;91m错误\033[0;1;91m：保存失败，请尝试选择新的保存位置。\033[0m")
-    root.destroy()
-    input("文件保存成功，按下回车键退出程序")
+        path = UI.askSaveFilePath(defaultType=".bcm", fileTypes=(("源码文件", ".bcm"), ("源码4文件", ".bcm4"), ("所有文件", ".*")))
+        if path == "":
+            if UI.askYesNot("要取消保存文件吗？"):
+                print("已取消保存文件")
+                return
+        else:
+            print(f"正在将源码保存到文件 \033[4m{path}\033[0m ……")
+            try:
+                with open(path, "w") as file:
+                    json.dump(sourceCode, file)
+                input("文件保存成功，按下回车键退出程序")
+                return
+            except Exception:
+                log(ERROR, "保存失败，请尝试选择新的保存位置。")
 
 if __name__ == "__main__":
-    main()
+    try:
+        main()
+    except:
+        pass
